+++ src/api/Controllers/FieldMeasurementController.cs
@@ -25,8 +25,8 @@ public class FieldMeasurementController : BdmsControllerBase<FieldMeasurement>
     public async Task<IEnumerable<FieldMeasurement>> GetAsync([FromQuery] int? boreholeId = null)
     {
         var fieldMeasurements = Context.FieldMeasurements
-            .Include(f => f.SampleType)
-            .Include(f => f.Parameter)
+            .Include(f => f.FieldMeasurementResults).ThenInclude(f => f.SampleType)
+            .Include(f => f.FieldMeasurementResults).ThenInclude(f => f.Parameter)
             .Include(f => f.Reliability)
             .Include(f => f.Casing)
             .ThenInclude(c => c.Completion)
@@ -40,22 +40,89 @@ public class FieldMeasurementController : BdmsControllerBase<FieldMeasurement>
         return await fieldMeasurements.ToListAsync().ConfigureAwait(false);
     }
 
-    /// <inheritdoc />
+    /// <summary>
+    /// Asynchronously creates the <paramref name="entity"/> specified.
+    /// </summary>
+    /// <param name="entity">The field measurement to create.</param>
+    [HttpPost]
     [Authorize(Policy = PolicyNames.Viewer)]
-    public override Task<ActionResult<FieldMeasurement>> EditAsync(FieldMeasurement entity)
-        => base.EditAsync(entity);
+    public override async Task<ActionResult<FieldMeasurement>> CreateAsync(FieldMeasurement entity) => await ProcessFieldMeasurement(entity).ConfigureAwait(false);
 
-    /// <inheritdoc />
+    /// <summary>
+    /// Asynchronously updates the <paramref name="entity"/> specified.
+    /// </summary>
+    /// <param name="entity">The field measurement to update.</param>
+    [HttpPut]
     [Authorize(Policy = PolicyNames.Viewer)]
-    public override Task<IActionResult> DeleteAsync(int id)
-        => base.DeleteAsync(id);
+    public override async Task<ActionResult<FieldMeasurement>> EditAsync(FieldMeasurement entity) => await ProcessFieldMeasurement(entity).ConfigureAwait(false);
 
-    /// <inheritdoc />
+    /// <summary>
+    /// Asynchronously deletes the field measurement with the specified <paramref name="id"/>.
+    /// </summary>
+    /// <param name="id">The id of the field measurement to delete.</param>
+    [HttpDelete]
     [Authorize(Policy = PolicyNames.Viewer)]
-    public override Task<ActionResult<FieldMeasurement>> CreateAsync(FieldMeasurement entity)
-        => base.CreateAsync(entity);
+    public override async Task<IActionResult> DeleteAsync(int id)
+    {
+        var fieldMeasurementToDelete = await Context.FindAsync(typeof(FieldMeasurement), id).ConfigureAwait(false);
+        if (fieldMeasurementToDelete == null)
+        {
+            return NotFound();
+        }
 
-    /// <inheritdoc />
+        // Check if associated borehole is locked
+        if (await BoreholeLockService.IsBoreholeLockedAsync(((FieldMeasurement)fieldMeasurementToDelete).BoreholeId, HttpContext.GetUserSubjectId()).ConfigureAwait(false))
+        {
+            return Problem("The borehole is locked by another user or you are missing permissions.");
+        }
+
+        Context.Remove(fieldMeasurementToDelete);
+        await Context.UpdateChangeInformationAndSaveChangesAsync(HttpContext).ConfigureAwait(false);
+        return Ok();
+    }
+
+    private async Task<ActionResult> ProcessFieldMeasurement(FieldMeasurement fieldMeasurement)
+    {
+        // Check if associated borehole is locked
+        if (await BoreholeLockService.IsBoreholeLockedAsync(fieldMeasurement.BoreholeId, HttpContext.GetUserSubjectId()).ConfigureAwait(false))
+        {
+            return Problem("The borehole is locked by another user or you are missing permissions.");
+        }
+
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(ModelState);
+        }
+
+        if (fieldMeasurement.Id != 0)
+        {
+            var fieldMeasurementToEdit = await Context.FieldMeasurements
+            .Include(f => f.FieldMeasurementResults)
+            .SingleOrDefaultAsync(f => f.Id == fieldMeasurement.Id).ConfigureAwait(false);
+
+            if (fieldMeasurementToEdit == null)
+            {
+                return NotFound();
+            }
+
+            UpdateFieldMeasurements(fieldMeasurement, fieldMeasurementToEdit);
+        }
+        else
+        {
+            await Context.AddAsync(fieldMeasurement).ConfigureAwait(false);
+        }
+
+        await Context.UpdateChangeInformationAndSaveChangesAsync(HttpContext).ConfigureAwait(false);
+        return Ok(fieldMeasurement);
+    }
+
+    private void UpdateFieldMeasurements(FieldMeasurement source, FieldMeasurement target)
+    {
+        Context.Entry(target).CurrentValues.SetValues(source);
+        target.FieldMeasurementResults = source.FieldMeasurementResults;
+    }
+
+        /// <inheritdoc />
     protected override Task<int?> GetBoreholeId(FieldMeasurement entity)
     {
         if (entity == null) return Task.FromResult<int?>(default);