+++ src/api/Models/FieldMeasurement.cs
@@ -8,30 +8,7 @@ namespace BDMS.Models;
 public class FieldMeasurement : Observation
 {
     /// <summary>
-    /// Gets or sets the <see cref="FieldMeasurement"/>'s sample type id.
+    /// Gets the <see cref="FieldMeasurementResult"/>s associated with the <see cref="FieldMeasurement"/>.
     /// </summary>
-    [Column("sample_type")]
-    public int SampleTypeId { get; set; }
-
-    /// <summary>
-    /// Gets or sets the <see cref="FieldMeasurement"/>'s sample type.
-    /// </summary>
-    public Codelist? SampleType { get; set; }
-
-    /// <summary>
-    /// Gets or sets the <see cref="FieldMeasurement"/>'s parameter id.
-    /// </summary>
-    [Column("parameter")]
-    public int ParameterId { get; set; }
-
-    /// <summary>
-    /// Gets or sets the <see cref="FieldMeasurement"/>'s parameter.
-    /// </summary>
-    public Codelist? Parameter { get; set; }
-
-    /// <summary>
-    /// Gets or sets the <see cref="FieldMeasurement"/>'s value.
-    /// </summary>
-    [Column("value")]
-    public double Value { get; set; }
+    public ICollection<FieldMeasurementResult>? FieldMeasurementResults { get; set; }
 }