+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 - Quality attribute for stratigraphy.
 - Support Amazon Cognito logout flow.
 - IsOpenBorehole option for instrumentation and backfill.
-- Show prompt for unsaved changes when switching data card or completion tabs.
+- Show prompt for unsaved changes when switching data card.
+- Support adding multiple field measurement results in `hydrogeology`.
 
 ### Changed